refactor: HomePage 폴더 변경
