+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import CategorySection from '@/components/HomePage/CategorySection';
+import CategorySection from '@/components/routes/home/CategorySection';
 
 export default function HomePage() {
   return <CategorySection />;
